Move track disambiguation for existing pages to dedicated file.
