+++ tests/test_common_utils/test_track_page_utils/test_wiiki_name_utils/test_track_disambiguation.py
@@ -3,30 +3,6 @@ import unittest
 from common_utils.track_page_utils.wiiki_name_utils import track_disambiguation as td
 
 class TestTrackDisambiguation(unittest.TestCase):
-    def test_read_authors_one(self):
-        page_name = "Ew Retro Track (notTarsa129)"
-        expected_set = {"notTarsa129"}
-        actual_set = td.read_authors(page_name)
-        self.assertSetEqual(expected_set, actual_set)
-
-    def test_read_authors_two(self):
-        page_name = "Ew Retro Track (notTarsa129 & randouser)"
-        expected_set = {"notTarsa129", "randouser"}
-        actual_set = td.read_authors(page_name)
-        self.assertSetEqual(expected_set, actual_set)
-
-    def test_read_authors_three(self):
-        page_name = "Ew Retro Track (notTarsa129, nottarsa & randouser)"
-        expected_set = {"notTarsa129", "nottarsa", "randouser"}
-        actual_set = td.read_authors(page_name)
-        self.assertSetEqual(expected_set, actual_set)
-
-    def test_read_authors_four(self):
-        page_name = "Ew Retro Track (notTarsa129, nottarsa, stillnottarsa & randouser)"
-        expected_set = {"notTarsa129", "nottarsa", "stillnottarsa", "randouser"}
-        actual_set = td.read_authors(page_name)
-        self.assertSetEqual(expected_set, actual_set)
-
     def get_page_from_name_authors(self):
         # Probably not good practice to not mock external API calls in unit tests.
         # This is why the name has been altered to NOT run as part of the test suite
